back/requirements.txt
fastapi
uvicorn
requests
feedparser
python-dateutil
beautifulsoup4
lxml
supabase
beautifulsoup4

